--- src/features/oauth2/index.ts
@@ -1,1 +1,0 @@
-export {default as useKakaoOAuthLogin} from './lib/useKakaoOAuthLogin';
